Move dump delaytx to separate class
